allow submission owners to download their own files
