+++ contest/urls.py
@@ -18,9 +18,7 @@ from django.conf import settings
 from django.conf.urls.static import static
 from django.contrib import admin
 from django.contrib.auth import views as auth_views
-from django.urls import include, path, re_path
-
-from contest import views
+from django.urls import include, path
 
 urlpatterns = [
     path('login', auth_views.LoginView.as_view(redirect_authenticated_user=True), name='login'),
@@ -41,7 +39,6 @@ urlpatterns = [
     path('', include('contests.urls')),
     path('schedule/', include('schedule.urls')),
     path('support/', include('support.urls')),
-    path('admin/', admin.site.urls),
-    re_path(r'^upload/', views.ProtectedUpload.as_view(), name='protected-upload')
+    path('admin/', admin.site.urls)
 ] + (static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT) +
      static(settings.STATIC_URL, document_root=settings.STATIC_ROOT))
